fix(ios): nudge Metal layer on theme change and expose bridge for Compose UI
When the system theme changes (light↔dark) while the app is backgrounded,
Compose recomposes with new colors after the existing scenePhase nudge has
already fired, leaving blank text. This adds a colorScheme observer with a
200ms delay so the glyph atlas rebuilds after theme recomposition.

Also introduces TextRenderingHelper — a commonMain bridge that lets Compose
UI trigger the Metal layer nudge from bottom tab clicks, giving users a
manual workaround if text goes blank.
