+++ homebase-common/src/commonMain/kotlin/id/homebase/core/TextRenderingHelper.kt
@@ -1,0 +1,17 @@
+package id.homebase.core
+
+interface TextRenderingNudger {
+    fun nudge()
+}
+
+object TextRenderingHelper {
+    private var _nudger: TextRenderingNudger? = null
+
+    fun register(nudger: TextRenderingNudger) {
+        _nudger = nudger
+    }
+
+    fun nudge() {
+        _nudger?.nudge()
+    }
+}
