+++ homebase-core/src/commonMain/kotlin/id/homebase/core/ui/navigation/AppNavHost.kt
@@ -63,6 +63,7 @@ import id.homebase.chat.editconversationgroup.EditConversationGroupScreen
 import id.homebase.chat.groupsettings.GroupSettingsScreen
 import id.homebase.chat.messageinfo.MessageInfoScreen
 import id.homebase.chat.selectmembers.SelectMembersScreen
+import id.homebase.core.TextRenderingHelper
 import id.homebase.core.navigation.ActiveConversation
 import id.homebase.core.notifications.NotificationNavigationEvent
 import id.homebase.core.permissions.PermissionStatus
@@ -326,6 +327,7 @@ fun AppNavHost(
                                 topLevelRoute.route::class
                             ) == true,
                             onClick = {
+                                TextRenderingHelper.nudge()
                                 if (topLevelRoute is TopLevelRoute.Vault) {
                                     openVault()
                                 } else {
@@ -359,6 +361,7 @@ fun AppNavHost(
                                 // label = { Text(stringResource(topLevelRoute.labelRes)) },
                                 selected = currentDestination?.hasRoute(topLevelRoute.route::class) == true,
                                 onClick = {
+                                    TextRenderingHelper.nudge()
                                     if (topLevelRoute is TopLevelRoute.Vault) {
                                         openVault()
                                     } else {
